tests: entity model: extract claim update tests to their own file
and deduplicate tests
